db: Add basic query and update helpers
The main appeal here is being able to pass in the values as separate
arguments rather than packed into tuple, which is particularly awkward
for a single value.  I think it's also nice for readability to have
functions that differentiate between writing and reading.

Though we'll probably move to SQLAlchemy anyway, in which case it
doesn't really matter...