schema: Add "failed" and "notneeded" status for tokens
Once verify_url() actually does something, it may fail to verify the
URL or skip the verification because it is already known.

Recording the failure is more important because /<dsid>/urls/<url>
will actually map that value to a status.  For known URLs, it won't
check the tokens table because the URL will already be present in the
dataset_urls table.