tests: Add fixture for generating dataset ID
The next commit will change the scope of the client fixture from
function to session.  That makes it more important to use different
dataset IDs across tests to reduce the chance of unintentional
interference when using the same database.

Add a fixture so that tests don't have to worry about coming up with
their one dataset IDs.