tests: Convert client to a session-scoped fixture
There's an interaction between the two tests that go through celery,
test_register_url_expired_token and test_register_url_expired_token,
which leads to the test that is executed second failing.  Each test
passes if executed on its own.

I don't have a good understanding of why this happens, but it seems
very likely to be an issue with the test setup rather than an
indication of an underlying problem with the application.

Make client a session-scoped fixture to sidestep this interaction.
The previous commit switched each test in test_datalad_urls.py over to
using a unique dsid, which should reduce the chances of tests
inadvertently interfering through the now shared database.