docs: Convert freeform API description to OpenAPI
This spec should include all the information that was in api.md, and
of course some details that weren't.  I'm new to OpenAPI, but I think
it's fairly complete, except for perhaps generic 404 responses.

This opens up the possibility of validating and testing with the spec,
or even using something like connexion to automatically map from the
spec to the endpoints.  It also allows us to use Swagger UI or any
other tool that understands OpenAPI (e.g., sphinxcontrib-openapi or
sphinxcontrib-redoc).

Another option worth considering is generating the spec from the code
and docstrings (e.g., using flasgger).  Even if we decide to go that
route, though, having the spec to work from will be helpful.