dataset_urls.urls.post: Change "url" key to "url_encoded"
In its POST response, urls() includes the encoded URL under the "url"
key.  To avoid confusion, we should either rename the key or return a
plain URL.  I think using an encoded URL makes sense: the request
already included the URL, so the caller is more likely to find the
encoded value useful.  Rename the key.