Avoid obsolete Celery settings
Celery reworked the settings names in v4.0 (mostly in terms of case
and prefixes) [1].  We're defining these settings along with
non-Celery ones, so I think it's helpful to continue to use the
"CELERY_" prefix.  Plus, if we continue to feed the settings through
flask's from_mapping(), we need to use uppercase because that method
discards lowercase names.

To continue using uppercase names that start with "CELERY_" while also
switching to the new names, process the settings with
celery.config_from_object(), which allows defining a namespace and
handles the case-conversion for uppercase namespaces.

[1] https://docs.celeryproject.org/en/stable/whatsnew-5.0.html