tests: Use in-memory database for client fixture
At some point I tried using an in-memory database for the tests, but
it failed.  However, that was before the SQLAlchemy switch and before
the latest Celery release, and it seems to work fine now.

Switch back to an in-memory database to avoid touching the disk for
each test that uses the client fixture.