Set DATALAD_REGISTRY_LOG_LEVEL in config module
create_app() calls _setup_logging() first thing, which checks the
DATALAD_REGISTRY_LOG_LEVEL environment variable and configures the
logger.  The idea was that the logger should be ready to go as soon as
possible, but it doesn't actually matter because no logging under the
datalad_registry namespace is done before the new location anyway.

Delay setting up the logging until after reading the configuration
sources so that DATALAD_REGISTRY_LOG_LEVEL can be handled in
datalad_registry.config, avoiding a special case.