factory: Read configuration from objects
create_app() sets up the configuration through a single call to
config.from_mapping().  This works right now for development, but a
clear separation between development and production configuration will
eventually be needed.

Make a config module with a DevelopmentConfig object and a placeholder
for the production config.  Leave the database URI handling in
create_app() because the default value uses the instance path.