tests: Share database connection across tests
Despite the tests passing at the moment, celery's database reference
can leak across the tests (see the example in gh-2).  For upcoming
tests that I'm working on, this mismatch causes failures.

I bet there's a proper way to set up a function-scoped fixture that
calls create_app() and doesn't have this problem, but after a good
amount of flailing, I haven't found it.

So, instead move the create_app() call, which brings along the setup
of the database and celery, into a session-scoped fixture.  Convert
app_instance(), the main fixture used by the tests, into a
function-scoped fixture that sits on top of the session-scoped
fixture, refreshing the tables before a test runs.

Closes #2.