views: Provide endpoint for listing known dataset IDs
I'm not sure if having a blueprint just for /datasets makes sense, but
at the same time I don't like the idea of moving all of the endpoints
under datasets/<dsid>/urls to the datasets.py blueprint.  In any case,
the arrangement is easy to adjust later, so go with the separate
blueprint for now.

Closes #5.