tests: Move dataset ID generation snippet to non-fixture function
Several tests rely on the dsid() fixture to get a dataset ID even when
they aren't actually creating a proper DataLad dataset.  An upcoming
test will use _multiple_ IDs in the same test, so it won't be able to
get the IDs from the fixture directly.  Expose a regular function that
it can use.