views: Extend URL endpoint with some collected information
collect_dataset_info() populates the remaining fields of URL, and, as
mentioned in gh-9, it might be useful to include some of some of that
information in 'GET /v1/datasets/{ds_id}/urls/{url_encoded}'.  It's
not obvious to me that this should simply dump all of the information,
including things used for internal tracking or things like "branches"
and "tags", which can be arbitrarily long.  Plus, the plan is to
collect a lot more information (e.g. metadata).

For now just add a few fields that seem to be the most useful.

Re: #9