tasks: Switch to DataLad for collecting information
collect_dataset_info() clones the URL with `git clone --mirror ...`.
That's convenient and easy but won't work once we start collecting
more information via DataLad.

Do a fairly minimal switch to using DataLad and a non-bare repo.  If
later desired/needed, some of the call_git() sites can be switched
over to GitRepo methods.

As indicated by the tests, the only change in collected information is
the inclusion of the remote HEAD symref in the list of branches.
That's not a useful addition, but it's also probably not worth
worrying about or filtering out at this point.