up: Wait for broker and celery node to be established
I haven't had any issues with using the 'up' script for interactive
testing, but, if this is going to be used for automated tests, we
should make sure things are ready before exiting.

The nc approach is based on a patch that John Wodder submitted for
DataLad's tools/ci/prep-travis-forssh.sh.