up-tests: Add 'up' wrapper that exits when test server is ready
Unlike the current tests, which make requests against test_client(),
the registry-submit and registry-announce-update tests will make
requests against 127.0.0.1:5000.  The 'up' script sets up everything
needed for these tests, but it finishes with `python -m flask run` and
sits in the foreground.  That's fine for interactive testing, but for
automated tests the top-level script needs to exit once 127.0.0.1:5000
is accessible.

Add an 'up' wrapper that is more useful for automated tests.