tasks: Add a directory level to dataset cache layout
collect_dataset_info() clones directly under
DATALAD_REGISTRY_DATASET_CACHE using the encoded URL as the directory
name.  That layout won't scale well to many datasets.

One option would be to add an intermediate level based on the first N
characters of the clone directory's name, but that doesn't have good
properties because URLs share a common prefix.  Instead add an
intermediate directory derived from the first three characters of the
dataset ID (with three chosen to follow what's done for RIA stores).
The leaf directory is still named according to the URL, since that is
the value that's unique.