Remove rate limit for `chk_url_to_update`
Celery counts revoked tasks due to expiration against
the rate limit. (This is possibly a bug in Celery).
Thus, rate limiting a task that is initiated with an
expiration can easily cause the task to be revoked due
to expiration. Additionally, `chk_url_to_update` tasks
are already "rate limited" by the setup of expiration and
the number of them being initiated in each run of the
`url_chk_dispatcher`, so this rate limit is not really
needed after all.