BF: switch to Earley parser, provide "parse_query" interface, separate out the rule for implicit AND
Apparently with "lalr" we didn't parse   :OP at all.  But switching to Earley
required to avoid "embedded" calling of the transformer so we had to come up
with "parse_query" adapter. Earley is said to be slower, but doesn't matter for
us -- our queries are small.

But then with Earley we started to get tests failing due to incorrect parsing
of implicit "AND" as prescribe in  "(AND|)".  It seems it works to extract that
implicit rule to be the last separate one -- then all tests pass etc.